Update navigation, refactor button component, add meta tags
This commit updates the navigation layout to include a Home link and changes the routing from 'task' to '/'. A new homepage is also created and SEO meta tags are added for the homepage and the 'create' page for better discoverability. The button component in the 'TaskControls' has been updated to use 'UiButton' for code consistency. Also, 'dev-host' script in package.json and webmanifest file are added. The task page is renamed to 'train'.